+++ config/config.exs
@@ -146,7 +146,7 @@ config :lightning, Lightning.Runtime.RuntimeManager, start: false
 
 config :lightning, LightningWeb.CollectionsController,
   default_stream_limit: 1_000,
-  max_database_limit: 5_000
+  max_database_limit: 500
 
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.